Build Huffman tree to compress a given string
We divide the string to individual character, then depending on the frequency of the characters, each will be bound to a different binary code (called Huffman Code). Combining the Huffman codes of all character, we'll get the complete Huffman string (all zeros and ones). In this solution, instead of using the built-in library heapq of python, I built a simple version of a priority queue, which basically just returns the most priority node (the one of the most frequency character) every time I evoke pop() method.